requirements.txt
fastapi
uvicorn[standard]
google-generativeai
groq
python-multipart
python-dotenv
pillow
aiofiles
aiohttp
